int8_t -> int32_t

+++ server/client.cpp
@@ -31,7 +31,7 @@ std::vector<std::uint8_t> Client::handOverInbox() {
 }
 
 void Client::send(const std::vector<std::uint8_t> &data) {
-  std::uint8_t length = data.size();
+  std::uint32_t length = data.size();
   bufferevent_write(_bufev, &length, 1);
   bufferevent_write(_bufev, &data[0], data.size());
 }
